UTTA.wiki/requirements.txt

accelerate>=0.20.0
accelerate>=0.20.0  # For faster training
backoff>=2.2.0  # For retry logic
beautifulsoup4>=4.12.0
bert-score>=0.3.13
bitsandbytes>=0.40.0
bitsandbytes>=0.40.0  # For quantization
colorama>=0.4.4
# Core dependencies
datasets>=2.10.0
datasets>=2.12.0  # HuggingFace datasets
# Document processing
dspy-ai>=2.0.0
# DSPy Optimization
ebooklib>=0.17.1
# Evaluation metrics
faiss-cpu>=1.7.0
# Fine-tuning dependencies
Flask>=2.3.0
Flask-Cors>=4.0.0
# HuggingFace Fine-tuning
jsonlines>=3.0.0
jsonlines>=3.1.0  # For JSONL file processing
langchain_openai>=0.0.1  # For OpenAI integration
# LLM and embedding dependencies
loguru>=0.6.0
matplotlib>=3.7.0
nltk>=3.6.0
nltk>=3.8.1
numpy>=1.20.0
numpy>=1.24.0 # Core dependencies
openai>=1.0.0
# OpenAI Fine-tuning
pandas>=1.3.0
pdfplumber>=0.7.0
peft>=0.4.0
peft>=0.5.0  # Parameter-Efficient Fine-Tuning
plotly>=5.14.0
PyPDF2>=2.0.0
python-docx>=0.8.11
python-dotenv>=0.19.0
pyyaml>=6.0
pyyaml>=6.0.1
requests>=2.26.0
rouge-score>=0.1.2
sacrebleu>=2.3.1
scikit-learn>=1.0.0  # For evaluation metrics
seaborn>=0.12.0
sentencepiece>=0.1.99
sentence-transformers>=2.2.0
tensorboard>=2.12.0 
# Text processing
tiktoken>=0.4.0
torch>=1.10.0
torch>=2.0.0
tqdm>=4.62.0
transformers>=4.20.0
transformers>=4.30.0
# Utilities
# Vector database
# Web application
Werkzeug>=2.3.0
